Added double precision atomicCAS-based atomic add for the gpu version: now GPU version is working with double precision. Should set up the build system to compile two sets of libraries

+++ helpers/base_types.py
@@ -1,22 +1,22 @@
 import numpy as np
 import ctypes
 global grid_type
-grid_type = np.complex64
+grid_type = np.complex128
 global psf_type
 psf_type = grid_type
 global fir_type
-fir_type = np.float32
+fir_type = np.float64
 global detaper_type
-detaper_type = np.float32
+detaper_type = np.float64
 global visibility_type
-visibility_type = np.complex64
+visibility_type = np.complex128
 global visibility_component_type
-visibility_component_type = np.float32
+visibility_component_type = np.float64
 global uvw_type
-uvw_type = np.float32
+uvw_type = np.float64
 global uvw_ctypes_convert_type
 uvw_ctypes_convert_type = ctypes.c_float if uvw_type == np.float32 else ctypes.c_double
 global reference_wavelength_type
-reference_wavelength_type = np.float32
+reference_wavelength_type = np.float64
 global weight_type
-weight_type = np.float32
+weight_type = np.float64
